refactor: メイン画面のコンポーネント移動
