+++ app/src/main/java/jp/co/yumemi/droidtraining/components/WeatherApp.kt
@@ -2,13 +2,7 @@ package jp.co.yumemi.droidtraining.components
 
 import WeatherFetchErrorDialog
 import android.content.res.Configuration.UI_MODE_NIGHT_YES
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Button
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
@@ -18,14 +12,8 @@ import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.RectangleShape
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
@@ -100,65 +88,6 @@ fun WeatherApp(
 
     if (updating) {
         LoadingOverlay()
-    }
-}
-
-@Composable
-fun WeatherInfo(weatherInfoData: WeatherInfoData, modifier: Modifier = Modifier) {
-    Column(modifier = modifier) {
-        Image(
-            painter = painterResource(id = weatherInfoData.icon),
-            contentDescription = weatherInfoData.weather.id,
-            contentScale = ContentScale.Crop,
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(RectangleShape),
-        )
-        WeatherTemperatureText(weather = weatherInfoData)
-    }
-}
-
-@Composable
-fun WeatherTemperatureText(weather: WeatherInfoData) {
-    Row {
-        Text(
-            text = "${weather.lowestTemperature}℃",
-            textAlign = TextAlign.Center,
-            modifier = Modifier.weight(1f),
-            color = Color.Blue,
-        )
-        Text(
-            text = "${weather.highestTemperature}℃",
-            textAlign = TextAlign.Center,
-            modifier = Modifier.weight(1f),
-            color = Color.Red,
-        )
-    }
-}
-
-@Composable
-fun ActionButtons(
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-    onReloadClick: () -> Unit = {},
-    onNextClick: () -> Unit = {},
-) {
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.SpaceEvenly,
-    ) {
-        Button(onClick = onReloadClick, enabled = enabled) {
-            Text(
-                text = stringResource(id = R.string.reload),
-                style = MaterialTheme.typography.labelMedium,
-            )
-        }
-        Button(onClick = onNextClick, enabled = enabled) {
-            Text(
-                text = stringResource(id = R.string.next),
-                style = MaterialTheme.typography.labelMedium,
-            )
-        }
     }
 }
